fix: refactor positionsUI
